Refactor: 일관성있는 이름을 위해 verifyUser.js 이름 변경
